requirements.txt
streamlit
pdfminer.six
python-docx
pandas
openai
dotenv